% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_print.R
\name{excel_print}
\alias{excel_print}
\title{Automatically save a collection of dataframes to a specified file directory in .xlsx format}
\usage{
excel_print(df_list, filepath, ggplots = FALSE)
}
\arguments{
\item{df_list}{Global environment list of dataframes}

\item{filepath}{Desired file path for the .xlsx files}

\item{ggplots}{Default FALSE - if TRUE, assumes df_list is a global environment list of ggplot objects, useful for 1:1 outputs with charts}
}
\value{
Returns nothing to the global environment
}
\description{
Automatically save a collection of dataframes to a specified file directory in .xlsx format
}
\examples{
\dontrun{
df1 <- data.frame(x = c(1,2,3), y = c("A", "B", "C"))
df2 <- data.frame(x = c(4,5,6), y = c("D", "E", "F"))

excel_print(mget(c("df1", "df2")), "Excel")

chart1 <- ggplot(mtcars, aes(x = mpg, y = disp)) + geom_point()
chart2 <- ggplot(mtcars, aes(x = mpg, y = hp)) + geom_point()

excel_print(mget(c("chart1", "chart2")), "Excel", ggplots = TRUE)

}
}
