% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwi_theme.R
\name{lwi_pal}
\alias{lwi_pal}
\title{Learning and Work Institute discrete colour palette}
\usage{
lwi_pal(fill = TRUE)
}
\arguments{
\item{fill}{Use the fill palette}
}
\description{
Learning and Work Institute discrete colour palette
}
\seealso{
Other colour lwi: 
\code{\link{scale_colour_lwi}()}
}
\concept{colour lwi}
