% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwi_theme.R
\name{lwi_theme}
\alias{lwi_theme}
\title{Learning and Work Institute ggplot theme}
\usage{
lwi_theme(base_size = 12, base_family = "sans", horizontal = TRUE)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{horizontal}{Default is TRUE for horizontal gridlines only. If FALSE, displays vertical gridlines, useful in conjunction with coord_flip()}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}
}
\description{
If Maven Pro Medium is installed (https://fonts.google.com/download?family=Maven\%20Pro)
and accessible by R through library(extrafont), the theme will use this font,
otherwise it will default to Arial.
}
