% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.R
\name{plot_print}
\alias{plot_print}
\title{Automatically save a collection of ggplot objects to a specified file directory as PNGs}
\usage{
plot_print(plot_list, filepath, width, height, res = 72, scaling = 1)
}
\arguments{
\item{plot_list}{List of ggplot objects from global environment}

\item{filepath}{Desired file path for the PNGs}

\item{width}{Width of the PNG output, in mm}

\item{height}{Height of the PNG output, in mm}

\item{res}{Resolution of the PNG output (default 72)}

\item{scaling}{Scaling factor of the PNG output (default 1)}
}
\value{
Returns nothing to the global environment
}
\description{
Automatically save a collection of ggplot objects to a specified file directory as PNGs
}
\examples{
\dontrun{
chart1 <- ggplot(mtcars, aes(x = mpg, y = disp)) + geom_point()
chart2 <- ggplot(mtcars, aes(x = mpg, y = hp)) + geom_point()

plot_print(mget(c("chart1", "chart2")), "Charts", width = 200, height = 200)
}
}
